add watchos test app for GDT and GDTCCT
