+++ GoogleDataTransport/GDTWatchOSTestApp/GDTWatchOSTestAppWatchKitApp/Assets.xcassets/AppIcon.appiconset/Contents.json
