+++ GoogleDataTransport/GDTWatchOSTestApp/GDTWatchOSTestAppWatchKitApp/Assets.xcassets/Contents.json
