+++ GoogleDataTransport/GDTWatchOSTestApp/GDTWatchOSTestAppWatchKitApp/Info.plist
