+++ GoogleDataTransport/GDTWatchOSTestApp/GDTWatchOSTestAppWatchKitExtension/Assets.xcassets/Complication.complicationset/Contents.json
