+++ GoogleDataTransport/GDTWatchOSTestApp/GDTWatchOSTestAppWatchKitExtension/Assets.xcassets/Complication.complicationset/Modular.imageset/Contents.json
