+++ GoogleDataTransport/GDTWatchOSTestApp/GDTWatchOSTestAppWatchKitExtension/Assets.xcassets/Complication.complicationset/Utilitarian.imageset/Contents.json
