+++ GoogleDataTransport/GDTWatchOSTestApp/GDTWatchOSTestAppWatchKitExtension/gdthelper.swift
