+++ GoogleDataTransportCCTSupport/GDTCCTWatchOSTestApp/GDTCCTWatchOSTestAppWatchKitApp/Assets.xcassets/AppIcon.appiconset/Contents.json
