+++ GoogleDataTransportCCTSupport/GDTCCTWatchOSTestApp/GDTCCTWatchOSTestAppWatchKitApp/Base.lproj/Interface.storyboard
@@ -9,31 +9,32 @@
         <!--Interface Controller-->
         <scene sceneID="aou-V4-d1y">
             <objects>
-                <controller id="AgC-eL-Hgc" customClass="InterfaceController" customModule="GDTWatchOSTestApp_WatchKit_App" customModuleProvider="target">
+                <controller id="AgC-eL-Hgc" customClass="InterfaceController" customModule="GDTCCTWatchOSTestAppWatchKitExtension">
                     <items>
-                        <button width="1" alignment="left" title="Generate data event" id="4Oh-eO-JpG">
+                        <label alignment="left" text="GDTCCT Test App" textAlignment="right" id="E4H-Lv-ipa"/>
+                        <button width="1" alignment="left" title="Generate data event" id="SDN-zp-3kg">
                             <connections>
-                                <action selector="generateDataEventWithSender:" destination="AgC-eL-Hgc" id="NUq-4p-3Xi"/>
+                                <action selector="generateDataEventWithSender:" destination="AgC-eL-Hgc" id="JUE-vH-PMu"/>
                             </connections>
                         </button>
-                        <button width="1" alignment="left" title="Generate telemetry event" id="GI0-zS-hCd">
+                        <button width="1" alignment="left" title="Generate telemetry event" id="bRh-1z-40x">
                             <connections>
-                                <action selector="generateTelemetryEventWithSender:" destination="AgC-eL-Hgc" id="daF-pb-Tpq"/>
+                                <action selector="generateTelemetryEventWithSender:" destination="AgC-eL-Hgc" id="HsT-yu-kQ3"/>
                             </connections>
                         </button>
-                        <button width="1" alignment="left" title="Generate high priority event (force uploads)" id="K1Q-Qs-F4h">
+                        <button width="1" alignment="left" title="Generate high priority event (force uploads)" id="bzu-1O-6S2">
                             <connections>
-                                <action selector="generateHighPriorityEventWithSender:" destination="AgC-eL-Hgc" id="vlS-Z8-MhS"/>
+                                <action selector="generateHighPriorityEventWithSender:" destination="AgC-eL-Hgc" id="IAi-r4-KCA"/>
                             </connections>
                         </button>
-                        <button width="1" alignment="left" title="Generate wifi only event" id="sbd-DF-6GA">
+                        <button width="1" alignment="left" title="Generate wifi only event" id="da7-Zz-zze">
                             <connections>
-                                <action selector="generateWifiOnlyEventWithSender:" destination="AgC-eL-Hgc" id="OiG-43-X22"/>
+                                <action selector="generateWifiOnlyEventWithSender:" destination="AgC-eL-Hgc" id="Bc5-M4-qez"/>
                             </connections>
                         </button>
-                        <button width="1" alignment="left" title="Generate daily event" id="i19-Ww-EpK">
+                        <button width="1" alignment="left" title="Generate daily event" id="3Nn-BQ-WRt">
                             <connections>
-                                <action selector="generateDailyEventWithSender:" destination="AgC-eL-Hgc" id="Izd-6h-qWV"/>
+                                <action selector="generateDailyEventWithSender:" destination="AgC-eL-Hgc" id="vw3-Pe-uuB"/>
                             </connections>
                         </button>
                     </items>
@@ -61,14 +62,14 @@
         <!--Notification Controller-->
         <scene sceneID="ZPc-GJ-vnh">
             <objects>
-                <controller id="4sK-HA-Art" customClass="NotificationController" customModule="GDTWatchOSTestApp_WatchKit_App" customModuleProvider="target"/>
+                <controller id="4sK-HA-Art" customClass="NotificationController" customModule="GDTCCTWatchOSTestAppWatchKitApp" customModuleProvider="target"/>
             </objects>
             <point key="canvasLocation" x="468" y="643"/>
         </scene>
         <!--Notification Controller-->
         <scene sceneID="Niz-AI-uX2">
             <objects>
-                <controller id="eXb-UN-Cd0" customClass="NotificationController" customModule="GDTWatchOSTestApp_WatchKit_App" customModuleProvider="target"/>
+                <controller id="eXb-UN-Cd0" customClass="NotificationController" customModule="GDTCCTWatchOSTestAppWatchKitApp" customModuleProvider="target"/>
             </objects>
             <point key="canvasLocation" x="468" y="345"/>
         </scene>
