+++ GoogleDataTransportCCTSupport/GDTCCTWatchOSTestApp/GDTCCTWatchOSTestAppWatchKitExtension/Assets.xcassets/Complication.complicationset/Extra Large.imageset/Contents.json
