+++ GoogleDataTransportCCTSupport/GDTCCTWatchOSTestApp/GDTCCTWatchOSTestAppWatchKitExtension/Assets.xcassets/Complication.complicationset/Graphic Corner.imageset/Contents.json
