+++ GoogleDataTransportCCTSupport/GDTCCTWatchOSTestApp/GDTCCTWatchOSTestAppWatchKitExtension/Assets.xcassets/Contents.json
