+++ GoogleDataTransportCCTSupport/GDTCCTWatchOSTestApp/GDTCCTWatchOSTestAppWatchKitExtension/ExtensionDelegate.swift
@@ -1,10 +1,18 @@
-//
-//  ExtensionDelegate.swift
-//  GDTCCTWatchOSTestApp WatchKit Extension
-//
-//  Created by Doudou Nan on 2/26/20.
-//  Copyright © 2020 Google. All rights reserved.
-//
+/*
+* Copyright 2020 Google
+*
+* Licensed under the Apache License, Version 2.0 (the "License");
+* you may not use this file except in compliance with the License.
+* You may obtain a copy of the License at
+*
+*      http://www.apache.org/licenses/LICENSE-2.0
+*
+* Unless required by applicable law or agreed to in writing, software
+* distributed under the License is distributed on an "AS IS" BASIS,
+* WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+* See the License for the specific language governing permissions and
+* limitations under the License.
+*/
 
 import WatchKit
