+++ GoogleDataTransportCCTSupport/GDTCCTWatchOSTestApp/Podfile
@@ -1,14 +1,14 @@
 use_frameworks!
 
-target 'GDTCCTWatchOSTestApp WatchKit App' do
+target 'GDTCCTWatchOSTestAppWatchKitApp' do
   platform :watchos, '6.0'
 
 end
 
-target 'GDTCCTWatchOSTestApp WatchKit Extension' do
+target 'GDTCCTWatchOSTestAppWatchKitExtension' do
   platform :watchos, '6.0'
 
-  # Pods for GDTCCTWatchOSTestApp WatchKit Extension
+  # Pods for GDTCCTWatchOSTestAppWatchKitExtension
   pod 'SwiftProtobuf', '~> 1.0'
   pod 'GoogleDataTransport', :path => '../../'
   pod 'GoogleDataTransportCCTSupport', :path => '../../'
